refactor(kernel): switch from u64 to usize

+++ kernel/src/acpi/facs.rs
@@ -6,9 +6,9 @@ use super::{Error, Header};
 
 pub const SIGNATURE: &[u8; 4] = b"FACS";
 
-static mut ADDR: u64 = 0;
+static mut ADDR: usize = 0;
 
-pub fn set_config(addr: u64) {
+pub fn set_config(addr: usize) {
     unsafe { ADDR = addr }
 }
