+++ kernel/src/drivers/storage/nvme/command/admin/identify/controller_data_structure.rs
@@ -645,7 +645,7 @@ struct PowerStateDescriptorData {
     qword3: u64,
 }
 
-pub fn handle(addr: u64) -> Result<(), crate::Error> {
+pub fn handle(addr: usize) -> Result<(), crate::Error> {
     Data::get_ref(addr).handle().delete()?;
     Ok(())
 }
