+++ kernel/src/drivers/storage/nvme/queue/mod.rs
@@ -25,11 +25,11 @@ impl Queue {
 
     pub fn init(
         &mut self,
-        doorbell_base: u64,
+        doorbell_base: usize,
         dstrd: u8,
         submission_size: u16,
         completion_size: u16,
-    ) -> Result<(u64, u64), crate::Error> {
+    ) -> Result<(usize, usize), crate::Error> {
         unsafe {
             self.id = ID_COUNTER;
             ID_COUNTER += 1;
@@ -37,11 +37,11 @@ impl Queue {
         Ok((
             self.submission.init(
                 submission_size,
-                doorbell_base + (2 * self.id as u64) * (1 << (2 + dstrd)),
+                doorbell_base + (2 * self.id as usize) * (1 << (2 + dstrd)),
             )?,
             self.completion.init(
                 completion_size,
-                doorbell_base + (2 * self.id as u64 + 1) * (1 << (2 + dstrd)),
+                doorbell_base + (2 * self.id as usize + 1) * (1 << (2 + dstrd)),
             )?,
         ))
     }
